requirements.txt
flask==3.0.2
requests==2.31.0
pytz==2023.3.post1
apscheduler==3.10.1
pytz==2023.3.post1
python-dotenv==1.0.0
configparser==6.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
Werkzeug